benchmark/01-react-tool-agent/requirements.txt
openai>=1.0.0,<2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
